Add utils and seL4 modules
The utils module contains mostly pure CakeML functions which perform
operations on words, integers and the like.

The seL4 module contains wrappers for some of the basic seL4 functions
for interacting with the kernel, which are required by the more
high-level invocations that are generated as part of CakeML libsel4.

Signed-off-by: Corey Lewis <corey.lewis@unsw.edu.au>